comments + variable name change
changed variable names to be consistent
we are using "frame" as in the context of MP3s in general, not as it is typically used in speech processing literature. a window is a series of frames.

changed variable names to be consistent across functions.

pitch_contour is time series of f0s returned from librosa.pyin

track is any time series audio, which includes pitch contours but also y directly from librosa.load